+++ R/utils.R
@@ -1,3 +1,49 @@
+#' Compute Odds From Probabilities
+#' 
+#' Compute odds from probabilities. The function is vectorized and
+#' can handle a vector of probabilities, e.g. remaining percent of participants
+#' as calculated by [compute_stats()]. 
+#' 
+#' @param p vector of probabilities. May not be larger than 1 or smaller than zero.
+#' @export
+#' 
+#' @returns Returns numerical vector of the same length as original input reflecting the odds.
+#' 
+#' @examples
+#' get_odds(0.7)
+#' @examples
+#' get_odds(c(0.7, 0.2))
+#' 
+#' 
+get_odds <- function(p){
+  if(!all(p <= 1 & p >= 0)) stop('Input is not a probability!')
+  p / (1-p)
+} 
+
+
+#' Compute Odds Ratio 
+#' 
+#' Computes odds ratio given two probabilities.
+#' In this package, the function can be used to compare the percentages of remaining
+#' participants between two conditions at a time.
+#' 
+#' 
+#' @param a numeric probability value between 0 and 1.
+#' @param b numeric probability value between 0 and 1.
+#' @export
+#' 
+#' @seealso [get_odds()], as this is the basis for calculation.
+#' 
+#' @returns Returns numerical vector of the same length as original input reflecting the Odds Ratio (OR).
+#' 
+#' @examples
+#' get_odds_ratio(0.7, 0.6)
+#' 
+get_odds_ratio <- function(a,b){
+  get_odds(a)/get_odds(b)
+}
+
+
 #' Calculate Steps for Uneven Data Points
 #'
 #' The `do_steps` function calculates steps for data points represented by numbers of questions from the original
@@ -63,4 +109,28 @@ do_steps <- function(x,
       list(x = xrep, y = yrep)  
     }
   }
+}
+
+
+#' Get Steps Data by Condition
+#' 
+#' The `get_steps_by_cond` function calculates steps data based on survival model results. 
+#' This utility function is used inside the [do_kpm()] function of `dropR`.
+#'
+#' @param sfit An object representing survival model results (e.g., from a Kaplan-Meier model).
+#' @param condition Optional. An experimental condition to include in the output data frame, defaults to `NULL`.
+#' 
+#' @seealso [do_kpm()]
+#'
+#' @returns Returns a data frame containing the steps data, including time, survival estimates, upper confidence bounds, and lower confidence bounds.
+#'
+#' @export
+get_steps_by_cond <- function(sfit, 
+                              condition = NULL){
+  f <- do_steps(sfit$time,sfit$surv)
+  u <- do_steps(sfit$time,sfit$upper)
+  l <- do_steps(sfit$time,sfit$lower)
+  dframe <- cbind(f,uppr = u$y, lwr = l$y)
+  if(!is.null(condition)) dframe$condition <- condition
+  dframe
 }