+++ NEWS.md
@@ -1,12 +1,11 @@
 # dropR (development version)
 
 - Install via `devtools::install_github("iscience-kn/dropR")`
-- Calculation for `add_dropout_idx()` and `compute_stats()` have been refined to allow more accuracy
-
 
 # dropR 1.0.4
 
-- Package will be updated on CRAN soon (as of April 2025), stay tuned!! 
+* Added new `do_print()` function for pretty output printing
+* Calculation for `add_dropout_idx()` and `compute_stats()` have been refined to allow more accuracy
 
 # dropR 1.0.3